frontend/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
flask-login==0.6.3
python-jose==3.3.0
flask-wtf==1.2.1
wtforms==3.1.1
email-validator==2.1.0.post1
flask-session==0.5.0